better script and plotter

+++ DM/src/main/java/ar/edu/itba/ss/Beeman.java
@@ -44,7 +44,7 @@ public class Beeman extends Oscillator {
     }
 
     @Override
-    public void firstStep(Double A, Integer omega) {
+    public void firstStep(Double A, Double omega) {
         setR(A * Math.sin(omega * getT()));
         this.setT(this.getT() + this.getDt());
     }
