+++ DM/src/main/java/ar/edu/itba/ss/Gear5.java
@@ -66,7 +66,7 @@ public class Gear5 extends Oscillator implements IntegrationMethod {
     }
 
     @Override
-    public void firstStep(Double A, Integer omega) {
+    public void firstStep(Double A, Double omega) {
         setR(A * Math.sin(omega * getT()));
     }
