+++ DM/src/main/java/ar/edu/itba/ss/Oscillator.java
@@ -30,7 +30,7 @@ public abstract class Oscillator {
 
     public abstract void coupledStep(Double prevY, Double nextY);
 
-    public abstract void firstStep(Double A, Integer omega);
+    public abstract void firstStep(Double A, Double omega);
 
     public abstract void lastStep();
